WIP: nanalogs set when no config file exists

+++ flyingpigeon/processes/wps_analogs_viewer.py
@@ -86,7 +86,7 @@ class AnalogsviewerProcess(WPSProcess):
             #Check if config file exists
             r = requests.get(configfile_with_path)
             if r.status_code != 404:
-                logger.debug('Config file exists on server address.')
+                logger.debug('Config file exists on server URL address.')
 
                 #Create dataframe and read in config file output by analogs detection process
                 #and stored on server URL address outputUrl_path()
@@ -96,26 +96,37 @@ class AnalogsviewerProcess(WPSProcess):
                 num_analogues = int(num_analogues.split( )[2])
 
             else:
-                logger.debug('Config file does not exist on server address. Read from local disk.')
+                logger.debug('Config file does not exist on server address. Check local disk.')
                
                 #Make config file name and get its path on local disk
                 configfile = 'config_' + analogs
+                ######configfile = 'configggg_' + analogs
                 p , name = os.path.split(os.path.realpath(analogs))
                 configfile_localAddress = os.path.join(p, configfile)
-                
-                #output_path (~/birdhouse/var/lib/pywps/outputs/flyingpigeon):
-                output_path = config.output_path()
 
-                #Create dataframe and read in config file stored on server URL address
-                dfC = pd.DataFrame()
-                dfC = pd.read_csv(configfile_localAddress, delimiter="none", skiprows=[15], index_col=0)
-              
-                num_analogues = dfC.index[9]
-                num_analogues = int(num_analogues.split( )[2])
+                #Check if config file exists
+                logger.debug('request: %s' % os.path.isfile(configfile_localAddress) )
 
-                #Copy config file to output_path (~/birdhouse/var/lib/pywps/outputs/flyingpigeon)
-                from shutil import copyfile
-                copyfile(configfile_localAddress, output_path + '/' + configfile)
+                if os.path.isfile(configfile_localAddress):
+                    logger.debug('Config file exists on local disk.')
+
+                    #output_path (~/birdhouse/var/lib/pywps/outputs/flyingpigeon):
+                    output_path = config.output_path()
+
+                    #Create dataframe and read in config file stored on server URL address
+                    dfC = pd.DataFrame()
+                    dfC = pd.read_csv(configfile_localAddress, delimiter="none", skiprows=[15], index_col=0)
+                  
+                    num_analogues = dfC.index[9]
+                    num_analogues = int(num_analogues.split( )[2])
+
+                    #Copy config file to output_path (~/birdhouse/var/lib/pywps/outputs/flyingpigeon)
+                    from shutil import copyfile
+                    copyfile(configfile_localAddress, output_path + '/' + configfile)
+                else:
+                    logger.debug('There is no config file on local disk. Set num_analogues = 20.')
+                    num_analogues = 20
+                    
 
         except Exception as e: 
             msg = 'failed to read number of analogues from config file %s ' % e
